// frontend/src/app/administracion-oee/datos-oee/models/datos.model.ts
export class DatosOee{
    idOeeInformacion: number;
    /* descripcionTipoDato: string; */
    /* descripcionOee: string; */
    descripcionOeeInformacion: string;
    oee: [];
    tipoDato: [];
    /* idTipoDato: string;
    idOee: string;
    descripcionOeeInformacion: string;
    estadoOeeInformacion: string; */
}